modules(nixos): Move KDE pkgs to desktop module
This also removes darkly input in the flake, to instead rely on nixpkgs one.

One important change is: Darkly now compiles against QT 6.10, while the current NixOS stable (25.05) has kdePackages compiled against 6.9.x, and because of this KDE fails to load the library.

To fix this, we override qtPackages for darkly package passing our current kdePackages.